+++ admin/Template/layout_1/LTR/material/full/customer_control.php
@@ -1,6 +1,15 @@
 <?php 
 
 include_once($_SERVER['DOCUMENT_ROOT'].'/'.'collage_canteen'.'/'.'config.php');
+$customersJson = file_get_contents($adminSources.'customers.json');
+$customers = json_decode($customersJson);
+foreach($customers as $customer){
+    $id[] = $customer->id;
+ }
+ sort($id);
+$lastIndex = count($id)-1;
+$highestId = $id[$lastIndex];
+$currentId = $highestId+1;
 $data = [
     "id"=>4,
     "name"=>$_POST['name'],
@@ -12,11 +21,13 @@ $data = [
     "passwoard"=>$_POST['passwoard'],
     "username"=>$_POST['username']
 ];
-$customersJson = file_get_contents($adminSources.'customers.json');
-$customers = json_decode($customersJson);
 $customers[] = (object)$data;
 $dataJson = json_encode($customers);
+if(file_exists($adminSources.'customers.json')){
 $result = file_put_contents($adminSources.'customers.json',$dataJson);
-
-echo "<h1 style = 'color:green;text-align:center;margin-top:100px;'>Customer is Added</h1>";
-echo "<a href='customers.php' style ='text-align:center;'>Customer List</a>";
+if($result){
+    location('customers.php');
+}
+}
+// echo "<h1 style = 'color:green;text-align:center;margin-top:100px;'>Customer is Added</h1>";
+// echo "<a href='customers.php' style ='text-align:center;'>Customer List</a>";
