+++ admin/Template/layout_1/LTR/material/full/product_controler.php
@@ -11,17 +11,26 @@ $e_sale ;
  if(isset($_POST['e-sale'])){
     $e_sale = $_POST;
  }else{
-    $e_sale = '';
+    $e_sale = null;
  }
 $outdoor;
  if(isset($_POST['outdoor'])){
     $outdoor = $_POST['outdoor'];
  }else{
-    $outdoor = '';
+    $outdoor = null;
  }
 $category = $_POST['category'];
+$productsJson = file_get_contents($adminSources.'products.json');
+$products = json_decode($productsJson);
+foreach($products as $product){
+   $id[] = $product->id;
+}
+sort($id);
+$lastIndex = count($id)-1;
+$highestId = $id[$lastIndex];
+$currentId = $highestId+1;
 $data = [
-    "id"=>9,
+    "id"=>$currentId,
     "name"=>$name,
     "type"=>$type,
     "description"=>$description,
@@ -33,10 +42,15 @@ $data = [
     "category"=>$category
 ];
 
-$productsJson = file_get_contents($adminSources.'products.json');
-$products = json_decode($productsJson);
+
 $products[] = (object)$data;
 $productData = json_encode($products);
+
+if(file_exists($adminSources.'products.json')){
 $result = file_put_contents($adminSources.'products.json',$productData);
+if($result){
+   location('products.php');
+}
+}
 echo "<h1 style='color:green;text-align:center;margin-top:80px;'>Your order is Added</h1>";
 echo "<a href='products.php'style='text-align:center;'>Your Product List</a>";
